backend/requirements.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0

# Authentication & Security
djangorestframework-simplejwt==5.3.0
python-decouple==3.8

# Database
psycopg2-binary==2.9.9

# API Documentation
drf-spectacular==0.26.5

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Code Quality
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Development
django-debug-toolbar==4.2.0
ipython==8.17.2

# Production
gunicorn==21.2.0
whitenoise==6.6.0
